Revert "Radamsa and ML RNN support in AFL (#762)"
This reverts commit c04fc20907e96392e509d020e8a7f4d130b15b1d.
