+++ src/python/bot/fuzzers/afl/launcher.py
@@ -253,22 +253,13 @@ class FuzzingStrategies(object):
   TIME_WITHOUT_FAST_CAL = 60 * 5
 
   # Strategy names to use in stats.
+  CORPUS_SUBSET_STRATEGY = 'strategy_corpus_subset'
   FAST_CAL_RANDOM_STRATEGY = 'strategy_fast_cal_random'
   FAST_CAL_MANUAL_STRATEGY = 'strategy_fast_cal_manual'
 
-  def __init__(self, target_path):
+  def __init__(self):
     strategy_pool = strategy_selection.generate_default_strategy_pool(
-        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=True)
-    # Select a generator to attempt to use for existing testcase mutations.
-    self.candidate_generator = engine_common.select_generator(
-        strategy_pool, target_path)
-    self.is_mutations_run = (
-        self.candidate_generator != engine_common.Generator.NONE)
-
-    # Generator that is actually used. Initialize to none, change if new
-    # testcase mutations are properly generated by the candidate generator.
-    self.generator_strategy = engine_common.Generator.NONE
-
+        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=False)
     self.use_corpus_subset = strategy_pool.do_strategy(
         strategy.CORPUS_SUBSET_STRATEGY)
 
@@ -326,21 +317,14 @@ class FuzzingStrategies(object):
   def print_strategies(self):
     """Print the strategies used for logging purposes."""
     fuzzing_strategies = []
-
-    if self.generator_strategy == engine_common.Generator.RADAMSA:
-      fuzzing_strategies.append('strategy_' +
-                                strategy.CORPUS_MUTATION_RADAMSA_STRATEGY.name)
-    elif self.generator_strategy == engine_common.Generator.ML_RNN:
-      fuzzing_strategies.append('strategy_' +
-                                strategy.CORPUS_MUTATION_ML_RNN_STRATEGY.name)
     if self.fast_cal == strategies.FastCal.RANDOM:
       fuzzing_strategies.append(self.FAST_CAL_RANDOM_STRATEGY)
     elif self.fast_cal == strategies.FastCal.MANUAL:
       fuzzing_strategies.append(self.FAST_CAL_MANUAL_STRATEGY)
 
     if self.use_corpus_subset:
-      fuzzing_strategies.append('strategy_{}_{}'.format(
-          strategy.CORPUS_SUBSET_STRATEGY.name, str(self.corpus_subset_size)))
+      fuzzing_strategies.append(self.CORPUS_SUBSET_STRATEGY + '_' +
+                                str(self.corpus_subset_size))
 
     print(engine_common.format_fuzzing_strategies(fuzzing_strategies))
 
@@ -555,20 +539,7 @@ class AflRunnerCommon(object):
     self._afl_input = None
     self._afl_output = None
 
-    self.strategies = FuzzingStrategies(target_path)
-    if self.strategies.is_mutations_run:
-      target_name = os.path.basename(self.target_path)
-      project_qualified_target_name = (
-          data_types.fuzz_target_project_qualified_name(utils.current_project(),
-                                                        target_name))
-      # Generate new testcase mutations according to candidate generator. If
-      # testcase mutations are properly generated, set generator strategy
-      # accordingly.
-      generator_used = engine_common.generate_new_testcase_mutations(
-          self.afl_input.input_directory, self.afl_input.input_directory,
-          project_qualified_target_name, self.strategies.candidate_generator)
-      if generator_used:
-        self.strategies.generator_strategy = self.strategies.candidate_generator
+    self.strategies = FuzzingStrategies()
 
     # Set this to None so we can tell if it has never been set or if it's just
     # empty.
@@ -979,8 +950,7 @@ class AflRunnerCommon(object):
     self._executable_path = self.afl_fuzz_path
 
     self.initial_max_total_time = (
-        get_fuzz_timeout(self.strategies.is_mutations_run) -
-        self.AFL_CLEAN_EXIT_TIME - self.SIGTERM_WAIT_TIME)
+        get_fuzz_timeout() - self.AFL_CLEAN_EXIT_TIME - self.SIGTERM_WAIT_TIME)
 
     assert self.initial_max_total_time > 0
 
@@ -1168,7 +1138,7 @@ class AflRunner(AflRunnerCommon, new_process.ProcessRunner):
 
 class MinijailAflRunner(AflRunnerCommon,
                         engine_common.MinijailEngineFuzzerRunner):
-  """Minijail AFL runner."""
+  """Minijail AFL runer."""
 
   def __init__(self,
                chroot,
@@ -1391,19 +1361,13 @@ def get_first_stacktrace(stderr_data):
   return stderr_data[:match.end()]
 
 
-def get_fuzz_timeout(is_mutations_run):
+def get_fuzz_timeout():
   """Get the maximum amount of time that should be spent fuzzing."""
   hard_timeout = engine_common.get_hard_timeout()
   merge_timeout = engine_common.get_merge_timeout(DEFAULT_MERGE_TIMEOUT)
   fuzz_timeout = hard_timeout - merge_timeout
-  mutations_timeout = engine_common.get_new_testcase_mutations_timeout()
-
-  if is_mutations_run:
-    fuzz_timeout -= mutations_timeout
-
-  assert fuzz_timeout > 0, (
-      'hard_timeout: %d merge_timeout: %d mutations_timeout: %d') % (
-          hard_timeout, merge_timeout, mutations_timeout)
+  assert fuzz_timeout > 0, 'hard_timeout: %d merge_timeout: %d' % (
+      hard_timeout, merge_timeout)
 
   return fuzz_timeout
