+++ src/python/fuzzing/strategy.py
@@ -82,8 +82,6 @@ LIBFUZZER_STRATEGY_LIST = [
 ]
 
 AFL_STRATEGY_LIST = [
-    CORPUS_MUTATION_RADAMSA_STRATEGY,
-    CORPUS_MUTATION_ML_RNN_STRATEGY,
     CORPUS_SUBSET_STRATEGY,
 ]
