+++ src/python/tests/core/bot/fuzzers/strategy_selection_test.py
@@ -25,9 +25,9 @@ from tests.test_libs import helpers as test_helpers
 from tests.test_libs import test_utils
 
 
-class TestDefaultStrategySelectionLibFuzzerPatched(unittest.TestCase):
+class TestRandomStrategySelectionGeneratorPatched(unittest.TestCase):
   """Tests whether program properly generates strategy pools for use by the
-  libFuzzer launcher."""
+  launcher."""
 
   def setUp(self):
     """Set up method for strategy pool generator tests with patch."""
@@ -58,9 +58,9 @@ class TestDefaultStrategySelectionLibFuzzerPatched(unittest.TestCase):
     self.assertTrue(strategy_pool.do_strategy(strategy.MUTATOR_PLUGIN_STRATEGY))
 
 
-class TestStrategySelectionLibFuzzerPatchless(unittest.TestCase):
+class TestStrategySelectionPatchless(unittest.TestCase):
   """Tests to see whether a default strategy pool is properly generated by the
-  file for the libFuzzer launcher."""
+  file."""
 
   def test_default_pool_generator(self):
     """Ensures that a call to generate_default_strategy_pool does not yield an
@@ -70,10 +70,9 @@ class TestStrategySelectionLibFuzzerPatchless(unittest.TestCase):
 
 
 @test_utils.with_cloud_emulators('datastore')
-class TestMultiArmedBanditStrategySelectionLibFuzzerPatch(unittest.TestCase):
+class TestMultiArmedBanditStrategySelectionPatch(unittest.TestCase):
   """Tests whether a multi armed bandit strategy pool is properly
-  generated according to the specified distribution for the libFuzzer
-  launcher."""
+  generated according to the specified distribution."""
 
   def setUp(self):
     """Put data in the local ndb table the tests to query from and set
@@ -131,9 +130,9 @@ class TestMultiArmedBanditStrategySelectionLibFuzzerPatch(unittest.TestCase):
 
 
 @test_utils.with_cloud_emulators('datastore')
-class TestMultiArmedBanditStrategySelectionLibFuzzer(unittest.TestCase):
+class TestMultiArmedBanditStrategySelection(unittest.TestCase):
   """Tests whether multi armed bandit strategy pool is properly generated
-  according to the specified distribution for the libFuzzer launcher.
+  according to the specified distribution.
 
   Deterministic tests. Only one strategy is put in the ndb table upon setup,
   so we know what the drawn strategy pool should be."""
@@ -228,176 +227,3 @@ class TestMultiArmedBanditStrategySelectionLibFuzzer(unittest.TestCase):
     self.assertFalse(strategy_pool.do_strategy(strategy.CORPUS_SUBSET_STRATEGY))
     self.assertFalse(strategy_pool.do_strategy(strategy.FORK_STRATEGY))
     self.assertTrue(strategy_pool.do_strategy(strategy.MUTATOR_PLUGIN_STRATEGY))
-
-
-class TestDefaultStrategySelectionAFLPatched(unittest.TestCase):
-  """Tests whether program properly generates strategy pools for use by the
-  AFL launcher."""
-
-  def setUp(self):
-    """Set up method for strategy pool generator tests with patch."""
-    test_helpers.patch_environ(self)
-    test_helpers.patch(self,
-                       ['bot.fuzzers.engine_common.decide_with_probability'])
-    self.mock.decide_with_probability.return_value = True
-
-  def test_default_pool_deterministic(self):
-    """Deterministically tests the default strategy pool generator."""
-    strategy_pool = strategy_selection.generate_default_strategy_pool(
-        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=True)
-
-    # Ml rnn and radamsa strategies are mutually exclusive. Because of how we
-    # patch, ml rnn will evaluate to false, however this depends on the
-    # implementation.
-    self.assertTrue(
-        strategy_pool.do_strategy(strategy.CORPUS_MUTATION_RADAMSA_STRATEGY))
-    self.assertFalse(
-        strategy_pool.do_strategy(strategy.CORPUS_MUTATION_ML_RNN_STRATEGY))
-    self.assertTrue(strategy_pool.do_strategy(strategy.CORPUS_SUBSET_STRATEGY))
-
-
-class TestStrategySelectionAFLPatchless(unittest.TestCase):
-  """Tests to see whether a default strategy pool is properly generated by the
-  file for the AFL launcher."""
-
-  def test_default_pool_generator(self):
-    """Ensures that a call to generate_default_strategy_pool does not yield an
-    exception. Deterministic behaviors are tested in the previous test."""
-    strategy_selection.generate_default_strategy_pool(
-        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=True)
-
-
-@test_utils.with_cloud_emulators('datastore')
-class TestMultiArmedBanditStrategySelectionAFLPatch(unittest.TestCase):
-  """Tests whether a multi armed bandit strategy pool is properly
-  generated according to the specified distribution for the AFL launcher."""
-
-  def setUp(self):
-    """Put data in the local ndb table the tests to query from and set
-    bandit selection environment variable."""
-    test_helpers.patch_environ(self)
-
-    data = []
-
-    strategy1 = data_types.FuzzStrategyProbability()
-    strategy1.strategy_name = 'corpus_mutations_ml_rnn,corpus_subset,'
-    strategy1.probability_medium_temperature = 0.33
-    strategy1.probability_high_temperature = 0.33
-    strategy1.probability_low_temperature = 0.33
-    data.append(strategy1)
-
-    strategy2 = data_types.FuzzStrategyProbability()
-    strategy2.strategy_name = ('corpus_mutations_radamsa,corpus_subset,')
-    strategy2.probability_medium_temperature = 0.34
-    strategy2.probability_high_temperature = 0.34
-    strategy2.probability_low_temperature = 0.34
-    data.append(strategy2)
-
-    strategy3 = data_types.FuzzStrategyProbability()
-    strategy3.strategy_name = ('corpus_subset,')
-    strategy3.probability_medium_temperature = 0.33
-    strategy3.probability_high_temperature = 0.33
-    strategy3.probability_low_temperature = 0.33
-    data.append(strategy3)
-    ndb.put_multi(data)
-
-    distribution = fuzz_task.get_strategy_distribution_from_ndb()
-
-    environment.set_value('USE_BANDIT_STRATEGY_SELECTION', True)
-    environment.set_value('STRATEGY_SELECTION_DISTRIBUTION', distribution)
-
-  def test_multi_armed_bandit_strategy_pool(self):
-    """Ensures a call to the multi armed bandit strategy selection function
-    doesn't yield an exception through any of the experimental paths."""
-    environment.set_value('STRATEGY_SELECTION_METHOD', 'default')
-    strategy_selection.generate_weighted_strategy_pool(
-        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=True)
-    environment.set_value('STRATEGY_SELECTION_METHOD',
-                          'multi_armed_bandit_medium')
-    strategy_selection.generate_weighted_strategy_pool(
-        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=True)
-    environment.set_value('STRATEGY_SELECTION_METHOD',
-                          'multi_armed_bandit_high')
-    strategy_selection.generate_weighted_strategy_pool(
-        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=True)
-    environment.set_value('STRATEGY_SELECTION_METHOD', 'multi_armed_bandit_low')
-    strategy_selection.generate_weighted_strategy_pool(
-        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=True)
-
-
-@test_utils.with_cloud_emulators('datastore')
-class TestMultiArmedBanditStrategySelectionAFL(unittest.TestCase):
-  """Tests whether multi armed bandit strategy pool is properly generated
-  according to the specified distribution for the AFL launcher.
-
-  Deterministic tests. Only one strategy is put in the ndb table upon setup,
-  so we know what the drawn strategy pool should be."""
-
-  def setUp(self):
-    """Put data in the local ndb table the tests to query from."""
-    test_helpers.patch_environ(self)
-    test_helpers.patch(self,
-                       ['bot.fuzzers.engine_common.decide_with_probability'])
-    self.mock.decide_with_probability.return_value = True
-
-    data = []
-
-    strategy1 = data_types.FuzzStrategyProbability()
-    strategy1.strategy_name = 'corpus_mutations_ml_rnn,corpus_subset,'
-    strategy1.probability_medium_temperature = 1
-    strategy1.probability_high_temperature = 1
-    strategy1.probability_low_temperature = 1
-    data.append(strategy1)
-    ndb.put_multi(data)
-
-    distribution = fuzz_task.get_strategy_distribution_from_ndb()
-
-    environment.set_value('USE_BANDIT_STRATEGY_SELECTION', True)
-    environment.set_value('STRATEGY_SELECTION_DISTRIBUTION', distribution)
-
-  def test_strategy_pool_medium_temperature(self):
-    """Tests whether a proper strategy pool is returned by the multi armed
-    bandit selection implementation with medium temperature.
-
-    Based on deterministic strategy selection. Mutator plugin is patched to
-    be included in our strategy pool."""
-    environment.set_value('STRATEGY_SELECTION_METHOD',
-                          'multi_armed_bandit_medium')
-    strategy_pool = strategy_selection.generate_weighted_strategy_pool(
-        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=True)
-    self.assertTrue(
-        strategy_pool.do_strategy(strategy.CORPUS_MUTATION_ML_RNN_STRATEGY))
-    self.assertFalse(
-        strategy_pool.do_strategy(strategy.CORPUS_MUTATION_RADAMSA_STRATEGY))
-    self.assertTrue(strategy_pool.do_strategy(strategy.CORPUS_SUBSET_STRATEGY))
-
-  def test_strategy_pool_high_temperature(self):
-    """Tests whether a proper strategy pool is returned by the multi armed
-    bandit selection implementation with high temperature.
-
-    Based on deterministic strategy selection. Mutator plugin is patched to
-    be included in our strategy pool."""
-    environment.set_value('STRATEGY_SELECTION_METHOD',
-                          'multi_armed_bandit_high')
-    strategy_pool = strategy_selection.generate_weighted_strategy_pool(
-        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=True)
-    self.assertTrue(
-        strategy_pool.do_strategy(strategy.CORPUS_MUTATION_ML_RNN_STRATEGY))
-    self.assertFalse(
-        strategy_pool.do_strategy(strategy.CORPUS_MUTATION_RADAMSA_STRATEGY))
-    self.assertTrue(strategy_pool.do_strategy(strategy.CORPUS_SUBSET_STRATEGY))
-
-  def test_strategy_pool_low_temperature(self):
-    """Tests whether a proper strategy pool is returned by the multi armed
-    bandit selection implementation with low temperature.
-
-    Based on deterministic strategy selection. Mutator plugin is patched to
-    be included in our strategy pool."""
-    environment.set_value('STRATEGY_SELECTION_METHOD', 'multi_armed_bandit_low')
-    strategy_pool = strategy_selection.generate_weighted_strategy_pool(
-        strategy_list=strategy.AFL_STRATEGY_LIST, use_generator=True)
-    self.assertTrue(
-        strategy_pool.do_strategy(strategy.CORPUS_MUTATION_ML_RNN_STRATEGY))
-    self.assertFalse(
-        strategy_pool.do_strategy(strategy.CORPUS_MUTATION_RADAMSA_STRATEGY))
-    self.assertTrue(strategy_pool.do_strategy(strategy.CORPUS_SUBSET_STRATEGY))
